+++ src/main/java/com/braintreegateway/CreditCard.java
@@ -9,6 +9,8 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 
+// NEXT_MAJOR_VERSION remove isVenmoSDK
+// The old venmo SDK integration has been deprecated
 public class CreditCard implements PaymentMethod {
     public static final String VALUE_YES = "Yes";
     public static final String VALUE_NO = "No";
@@ -179,6 +181,7 @@ public class CreditCard implements PaymentMethod {
     private String expirationMonth;
     private String expirationYear;
     private boolean isDefault;
+    @Deprecated
     private boolean isVenmoSdk;
     private boolean isExpired;
     private boolean isNetworkTokenized;
@@ -378,6 +381,11 @@ public class CreditCard implements PaymentMethod {
         return isDefault;
     }
 
+    //NEXT_MAJOR_VERSION remove this method
+    /**
+     * @deprecated - The Venmo SDK integration is Unsupported. Please update your integration to use Pay with Venmo instead
+    */
+    @Deprecated
     public boolean isVenmoSdk() {
       return isVenmoSdk;
     }