+++ src/main/java/com/braintreegateway/TransactionRequest.java
@@ -11,6 +11,8 @@ import java.util.Map;
 /**
  * Provides a fluent interface to build up requests around {@link Transaction Transactions}.
  */
+// NEXT_MAJOR_VERSION remove venmoSdkPaymentMethodCode
+// The old venmo SDK integration has been deprecated
 public class TransactionRequest extends Request {
     private BigDecimal amount;
     private TransactionAddressRequest billingAddressRequest;
@@ -46,6 +48,7 @@ public class TransactionRequest extends Request {
     private BigDecimal discountAmount;
     private String shipsFromPostalCode;
     private Type type;
+    @Deprecated
     private String venmoSdkPaymentMethodCode;
     private String paymentMethodNonce;
     private BigDecimal serviceFeeAmount;
@@ -250,6 +253,10 @@ public class TransactionRequest extends Request {
         return this;
     }
 
+    //NEXT_MAJOR_VERSION remove this method
+    /**
+     * @deprecated - The Venmo SDK integration is Unsupported. Please update your integration to use Pay with Venmo instead
+    */
     public TransactionRequest shipsFromPostalCode(String shipsFromPostalCode) {
         this.shipsFromPostalCode = shipsFromPostalCode;
         return this;
@@ -361,6 +368,8 @@ public class TransactionRequest extends Request {
         return this.installments;
     }
 
+    // NEXT_MAJOR_VERSION remove venmoSdkPaymentMethodCode
+    // The old venmo SDK integration has been deprecated
     protected RequestBuilder buildRequest(String root) {
         RequestBuilder builder = new RequestBuilder(root)
             .addElement("amount", amount)