+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 * Deprecate `threeDSecureToken` in favor of `threeDSecureAuthenticationId`
 * Add `externalNetworkTokenizationEnrollmentId` to `PaymentMethodGrantRequest`
+* Deprecate `isVenmoSdk`, `venmoSdkSession`, and `venmoSdkPaymentMethodCode`
 
 ## 3.29.0
 * Add methods to `UsBankAccount`